backend/app/requirements.txt
fastapi==0.115.12
psycopg2-binary==2.9.10
uvicorn==0.34.2
